+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -85,6 +85,12 @@ Improvements to clang-tidy
   Finds ``memset()`` calls with potential mistakes in their arguments.
   Replaces and extends the ``google-runtime-memset`` check.
 
+- New `bugprone-undefined-memory-manipulation
+  <http://clang.llvm.org/extra/clang-tidy/checks/bugprone-undefined-memory-manipulation.html>`_ check
+
+  Finds calls of memory manipulation functions ``memset()``, ``memcpy()`` and
+  ``memmove()`` on not TriviallyCopyable objects resulting in undefined behavior.
+
 - New `cert-dcl21-cpp
   <http://clang.llvm.org/extra/clang-tidy/checks/cert-dcl21-cpp.html>`_ check